frontend/requirements.txt
streamlit==1.36.0
streamlit_pdf_viewer==0.0.14
pandas==2.2.2
plotly==5.22.0
azure-storage-blob==12.20.0
azure-cosmos==4.7.0
python-dotenv==1.0.1
streamlit-authenticator=0.3.3